% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stitching_site.R
\name{prepareSite}
\alias{prepareSite}
\title{Prepares the site}
\usage{
prepareSite(ymlFile, htmlFilePath = getwd())
}
\arguments{
\item{ymlFile}{path to yml file which defines the site}

\item{htmlFilePath}{directory path to all html files ( default current working directory)}
}
\value{
Generates html files with navigation bar in the output_dir as defined in yml file
}
\description{
Generates navigation bar as defined in ymlFile , inserts the navigation bar of every local html files defined in yml file and available inside htmlFilePath.
}
\examples{
prepareSite(ymlFile = "site.yml")
}
